+++ src/main/java/app/AppBuilder.java
@@ -5,9 +5,9 @@ import entity.CountryFactory;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.save_country.SaveCountryViewModel;
 import interface_adapter.profile.ProfileViewModel;
-import interface_adapter.search.ByLanguage.SearchByLanguageController;
-import interface_adapter.search.ByLanguage.SearchByLanguagePresenter;
-import interface_adapter.search.ByLanguage.SearchByLanguageViewModel;
+import interface_adapter.search.by_language.SearchByLanguageController;
+import interface_adapter.search.by_language.SearchByLanguagePresenter;
+import interface_adapter.search.by_language.SearchByLanguageViewModel;
 import use_case.search.by_language.SearchByLanguageCountryDataAccessInterface;
 import use_case.search.by_language.SearchByLanguageInputBoundary;
 import use_case.search.by_language.SearchByLanguageInteractor;
